+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- `Metadata` dataclass for `Parameter` metadata.
+- Generic `Metadata` system for BayBE objects containing description, unit, and additional information.
 
 ### Fixed
 - The Python version specifier now also allows patch versions of Python 3.13